chore: rename TokenService and it methods to ERC20
