+++ src/controllers/token/erc20/get.action.ts
@@ -1,8 +1,8 @@
 import { Request, Response } from 'express';
-import TokenService from '@/services/TokenService';
+import ERC20Service from '@/services/ERC20Service';
 
-export const getERC20TokenById = async (req: Request, res: Response) => {
+export const getById = async (req: Request, res: Response) => {
     const id = req.params['id'];
-    const token = await TokenService.getERC20TokenById(id);
+    const token = await ERC20Service.getById(id);
     return res.send({ token: token.toJSON() });
 };
