+++ src/controllers/token/erc20/get.all.action.ts
@@ -1,7 +1,7 @@
-import TokenService from '@/services/TokenService';
+import ERC20Service from '@/services/ERC20Service';
 import { Request, Response } from 'express';
 
 export const getAllERC20Token = async (req: Request, res: Response) => {
-    const tokens = await TokenService.getAllERC20TokenBySub(req.user.sub);
+    const tokens = await ERC20Service.getAll(req.user.sub);
     return res.send({ tokens: tokens.map((token) => token.toJSON()) });
 };
