+++ src/controllers/token/erc20/post.transfer.ts
@@ -1,4 +1,4 @@
 import { Request, Response } from 'express';
-import TokenService from '@/services/TokenService';
+import ERC20Service from '@/services/ERC20Service';
 
 const transferToken = async (req: Request, res: Response) => {};
